Handle data received as plain text
This allows us to avoid CORS preflight requests if we want to, at the expense
of sending the correct content-type. Sending data as `application/json` in this
scenario will trigger the browser to send an OPTIONS request before the POST,
but sending as `text/plain` avoids it. Take your pick.